refactor column utility code to model_column_util
